requirements.txt
click==7.0
influxdb==5.2.1
PyYAML==3.13
pyzmq==17.1.2
schema==0.6.8
